Parse file lists more thoroughly
Implemented a basic parser instead of using luas string substitution to
deal with file lists. In theory we can now deal with : and ; being in
file names. It now parses two styles of entries:
- `{/path/to/file}`
- `{name:/path/to/file:ext}`
and can do a list of entires with mixed styles. Currently the first one
is opened instead of actually dealing with the list of entries.